// src/types/api.ts
export type Notification = {
  __typename: "Notification",
  id: string,
  userId: string,
  message: string,
  link?: string | null,
  relatedUserId: string,
  relatedUser: User,
  primaryPicture?: S3Object | null,
  secondaryPicture?: S3Object | null,
  type: NotificationTypes,
  createdAt: string,
  updatedAt: string,
  userRelatedNotificationsId?: string | null,
  owner?: string | null,
};

export type User = {
  __typename: "User",
  id: string,
  personFirstName?: string | null,
  personLastName?: string | null,
  personFullName?: string | null,
  venueName?: string | null,
  venueSearchName?: string | null,
  venuePhone?: string | null,
  venueAddressCountry?: string | null,
  venueAddressCity?: string | null,
  venueAddressStreet?: string | null,
  venueAddressNumber?: string | null,
  venueAddressState?: string | null,
  venueAddressGeo?: GeoPoint | null,
  venueCalculatedRating?: number | null,
  venueMenu?: S3Object | null,
  venueCheckinsCount?: number | null,
  venueWebsite?: string | null,
  venueHours?: string | null,
  brandName?: string | null,
  brandSearchName?: string | null,
  brandDescription?: string | null,
  brandLogo?: S3Object | null,
  brandCoverImage?: S3Object | null,
  brandWebsite?: string | null,
  brandCountry?: string | null,
  brandFoundedYear?: number | null,
  brandStory?: string | null,
  brandId?: string | null,
  userType: UserType,
  username: string,
  profilePicture?: S3Object | null,
  profilePictureKey?: string | null,
  profilePictureModerationStatus?: ModerationStatus | null,
  coverPicture?: S3Object | null,
  coverPictureKey?: string | null,
  coverPictureModerationStatus?: ModerationStatus | null,
  bio?: string | null,
  notificationSettings?:  Array<Settings | null > | null,
  securitySettings?:  Array<Settings | null > | null,
  deleted?: boolean | null,
  archived?: boolean | null,
  isMyCollectionPublic?: boolean | null,
  whiskeys?: ModelUserWhiskeysConnection | null,
  reviews?: ModelReviewConnection | null,
  likedPosts?: ModelUserLikesConnection | null,
  posts?: ModelPostConnection | null,
  comments?: ModelCommentConnection | null,
  wishList?: ModelUserWishListWhiskeysConnection | null,
  trendingWhiskeys?: ModelUserTrendingWhiskeysConnection | null,
  followers?: Array< string | null > | null,
  following?: Array< string | null > | null,
  expoTokens?: Array< string | null > | null,
  relatedNotifications?: ModelNotificationConnection | null,
  reports?: ModelReportConnection | null,
  rewards?: ModelUserRewardConnection | null,
  isOnRewards?: boolean | null,
  favoriteGuides?: ModelUserFavoriteGuidesConnection | null,
  myPours?: ModelUserPoursConnection | null,
  brandWhiskeys?: ModelWhiskeyConnection | null,
  brandVenues?: ModelUserConnection | null,
  cmsUsers?: ModelCMSUserBrandConnection | null,
  blockedUsers?: Array< string | null > | null,
  toBeRedeemed?: boolean | null,
  externalId?: string | null,
  clubMembers?: ModelClubMemberConnection | null,
  clubRequestsCreated?: ModelClubRequestConnection | null,
  dmPrivacySetting?: DMPrivacySetting | null,
  conversationParticipants?: ModelConversationParticipantConnection | null,
  messages?: ModelMessageConnection | null,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type GeoPoint = {
  __typename: "GeoPoint",
  lat: number,
  lon: number,
};

export type S3Object = {
  __typename: "S3Object",
  bucket: string,
  key: string,
  region: string,
  width?: number | null,
  height?: number | null,
};

export enum UserType {
  VENUE = "VENUE",
  PERSON = "PERSON",
  BRAND = "BRAND",
}


export enum ModerationStatus {
  PENDING = "PENDING",
  APPROVED = "APPROVED",
  REJECTED = "REJECTED",
  ERROR = "ERROR",
}


export type Settings = {
  __typename: "Settings",
  name: string,
  description: string,
  value: string,
};

export type ModelUserWhiskeysConnection = {
  __typename: "ModelUserWhiskeysConnection",
  items:  Array<UserWhiskeys | null >,
  nextToken?: string | null,
};

export type UserWhiskeys = {
  __typename: "UserWhiskeys",
  userId: string,
  user?: User | null,
  whiskeyId: string,
  whiskeyFullName?: string | null,
  whiskey?: Whiskey | null,
  geo?: GeoPoint | null,
  proof?: number | null,
  proofType?: ProofType | null,
  warehouse?: string | null,
  purchaseYear?: string | null,
  batch?: string | null,
  rick?: string | null,
  barrel?: string | null,
  bottle?: string | null,
  notes?: string | null,
  singleBarrel?: boolean | null,
  storePick?: string | null,
  age?: string | null,
  style?: string | null,
  whiskeyType?: Array< WhiskeyType | null > | null,
  archived?: boolean | null,
  id: string,
  createdAt: string,
  updatedAt: string,
  userWhiskeysId?: string | null,
  whiskeyUsersId?: string | null,
  owner?: string | null,
};
// ...
export type ModelUserWishListWhiskeysConnection = {
  __typename: "ModelUserWishListWhiskeysConnection",
  items:  Array<UserWishListWhiskeys | null >,
  nextToken?: string | null,
};

export type UserWishListWhiskeys = {
  __typename: "UserWishListWhiskeys",
  id: string,
  userId: string,
  whiskeyId: string,
  user: User,
  whiskey: Whiskey,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type ModelUserTrendingWhiskeysConnection = {
  __typename: "ModelUserTrendingWhiskeysConnection",
  items:  Array<UserTrendingWhiskeys | null >,
  nextToken?: string | null,
};

export type UserTrendingWhiskeys = {
  __typename: "UserTrendingWhiskeys",
  userId: string,
  user?: User | null,
  whiskeyId: string,
  whiskey?: Whiskey | null,
  count?: number | null,
  id: string,
  createdAt: string,
  updatedAt: string,
  userTrendingWhiskeysId?: string | null,
  whiskeyUserTrendingWhiskeysId?: string | null,
  owner?: string | null,
};

export type ModelReviewConnection = {
  __typename: "ModelReviewConnection",
  items:  Array<Review | null >,
  nextToken?: string | null,
};

export type Review = {
  __typename: "Review",
  id: string,
  title?: string | null,
  description?: string | null,
  rating: number,
  recommendationTags?: Array< ReviewRecommendationTags | null > | null,
  whiskey?: Whiskey | null,
  whiskeyId: string,
  userId: string,
  user?: User | null,
  specialistReview?: number | null,
  specialistName?: string | null,
  specialistImage?: S3Object | null,
  createdAt: string,
  updatedAt: string,
  userReviewsId?: string | null,
  whiskeyReviewsId?: string | null,
  owner?: string | null,
};

export enum ReviewRecommendationTags {
  OAKY = "OAKY",
  FRUITY = "FRUITY",
  NUTTY = "NUTTY",
  BRINY = "BRINY",
  BUTTERY = "BUTTERY",
  SPICY = "SPICY",
  SMOKY = "SMOKY",
  EARTHY = "EARTHY",
  HERBAL = "HERBAL",
  RICH = "RICH",
  SILKY = "SILKY",
}


export type ModelUserPoursConnection = {
  __typename: "ModelUserPoursConnection",
  items:  Array<UserPours | null >,
  nextToken?: string | null,
};

export type UserPours = {
  __typename: "UserPours",
  userId: string,
  user?: User | null,
  whiskeyFullName?: string | null,
  count?: number | null,
  whiskeyId: string,
  whiskey?: Whiskey | null,
  id: string,
  createdAt: string,
  updatedAt: string,
  userMyPoursId?: string | null,
  whiskeyUserPoursId?: string | null,
  owner?: string | null,
};
// ...
export type ModelPostConnection = {
  __typename: "ModelPostConnection",
  items:  Array<Post | null >,
  nextToken?: string | null,
};

export type Post = {
  __typename: "Post",
  id: string,
  description?: string | null,
  authorId: string,
  locationId?: string | null,
  author?: User | null,
  photo?: S3Object | null,
  photoKey?: string | null,
  photoModerationStatus?: ModerationStatus | null,
  photos?:  Array<S3Object | null > | null,
  photoKeys?: Array< string | null > | null,
  photoWidths?: Array< number | null > | null,
  photoHeights?: Array< number | null > | null,
  comments?: ModelCommentConnection | null,
  likes?: ModelUserLikesConnection | null,
  likesCount: number,
  title?: string | null,
  references?:  Array<Reference | null > | null,
  inlineTags?:  Array<InlineTag | null > | null,
  sharedPostId?: string | null,
  sharedPost?: Post | null,
  shareComment?: string | null,
  sharesCount?: number | null,
  shares?: ModelPostConnection | null,
  createdAt: string,
  clubId?: string | null,
  clubIsPrivate?: boolean | null,
  club?: Club | null,
  updatedAt: string,
  userPostsId?: string | null,
  postSharesId?: string | null,
  clubPostsId?: string | null,
  owner?: string | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items:  Array<Comment | null >,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  text: string,
  authorId: string,
  author?: User | null,
  postId: string,
  post?: Post | null,
  createdAt: string,
  updatedAt: string,
  userCommentsId?: string | null,
  postCommentsId?: string | null,
  owner?: string | null,
};

export type ModelUserLikesConnection = {
  __typename: "ModelUserLikesConnection",
  items:  Array<UserLikes | null >,
  nextToken?: string | null,
};

export type UserLikes = {
  __typename: "UserLikes",
  id: string,
  userId: string,
  postId: string,
  user: User,
  post: Post,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};
// ...
export type ModelClubMemberConnection = {
  __typename: "ModelClubMemberConnection",
  items:  Array<ClubMember | null >,
  nextToken?: string | null,
};
// ...
export type ModelCMSUserBrandConnection = {
  __typename: "ModelCMSUserBrandConnection",
  items:  Array<CMSUserBrand | null >,
  nextToken?: string | null,
};

export type CMSUserBrand = {
  __typename: "CMSUserBrand",
  id: string,
  cmsUserId: string,
  cmsUser?: CMSUser | null,
  brandUserId: string,
  brandUser?: User | null,
  assignedAt: string,
  assignedBy?: string | null,
  createdAt: string,
  updatedAt: string,
  userCmsUsersId?: string | null,
  cMSUserBrandsId?: string | null,
};

export type ModelClubRequestConnection = {
  __typename: "ModelClubRequestConnection",
  items:  Array<ClubRequest | null >,
  nextToken?: string | null,
};

export type ModelNotificationConnection = {
  __typename: "ModelNotificationConnection",
  items:  Array<Notification | null >,
  nextToken?: string | null,
};

export type ModelReportConnection = {
  __typename: "ModelReportConnection",
  items:  Array<Report | null >,
  nextToken?: string | null,
};

export type Report = {
  __typename: "Report",
  id: string,
  reason: string,
  description?: string | null,
  reportedUserId: string,
  reportedUser?: User | null,
  contentId: string,
  contentType: ReportContentType,
  createdAt: string,
  updatedAt: string,
  userReportsId?: string | null,
};

export enum ReportContentType {
  POST = "POST",
  REVIEW = "REVIEW",
  COMMENT = "COMMENT",
}


export type ModelUserRewardConnection = {
  __typename: "ModelUserRewardConnection",
  items:  Array<UserReward | null >,
  nextToken?: string | null,
};

export type UserReward = {
  __typename: "UserReward",
  id: string,
  userId: string,
  user?: User | null,
  rewardId: string,
  reward?: Reward | null,
  score: number,
  lastScoreUpdate: string,
  isAlreadyViewed?: boolean | null,
  isRedeemed?: boolean | null,
  isCompleted?: boolean | null,
  trackingCode?: string | null,
  service?: string | null,
  size?: string | null,
  model?: string | null,
  fullName?: string | null,
  email?: string | null,
  address?: string | null,
  city?: string | null,
  state?: string | null,
  zipcode?: string | null,
  createdAt: string,
  updatedAt: string,
  userRewardsId?: string | null,
  rewardUsersId?: string | null,
  owner?: string | null,
};
// ...
export type ModelUserFavoriteGuidesConnection = {
  __typename: "ModelUserFavoriteGuidesConnection",
  items:  Array<UserFavoriteGuides | null >,
  nextToken?: string | null,
};

export type UserFavoriteGuides = {
  __typename: "UserFavoriteGuides",
  id: string,
  userId: string,
  guidesId: string,
  user: User,
  guides: Guides,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type Guides = {
  __typename: "Guides",
  id: string,
  title: string,
  subtitle: string,
  coverPhoto: S3Object,
  tag: string,
  body: Array< string | null >,
  photos:  Array<S3Object | null >,
  users?: ModelUserFavoriteGuidesConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelWhiskeyConnection = {
  __typename: "ModelWhiskeyConnection",
  items:  Array<Whiskey | null >,
  nextToken?: string | null,
};

export enum DMPrivacySetting {
  EVERYONE = "EVERYONE",
  FOLLOWING = "FOLLOWING",
  NONE = "NONE",
}


export type ModelConversationParticipantConnection = {
  __typename: "ModelConversationParticipantConnection",
  items:  Array<ConversationParticipant | null >,
  nextToken?: string | null,
};
// ...
export type ModelMessageConnection = {
  __typename: "ModelMessageConnection",
  items:  Array<Message | null >,
  nextToken?: string | null,
};
// ...
export type ModelUserConnection = {
  __typename: "ModelUserConnection",
  items:  Array<User | null >,
  nextToken?: string | null,
};